Landing: Use on-chain NFT-Metadata
On the landing page, the actual on-chain metadata shall be shown. This
is fetched via an NFT's token_uri, which is available through the
asset-by-sip endpoint.

This displays and fetches all data through the on-chain tokenURI, which
is conveniently provided through metaanchor-api.

In many settings the instance displaying the landing page may also serve
the metadata and NFT-Assets, but especially if IPFS or other customized
settings are used, which overwrite on-chain TokenURI / baseURI, this
breaks.

This also shows and embeds OpenSea links, which is provided by
MetaAnchor-API v0.2.3 and newer