Define /srv/data Mountpoint for nftdata and .next/cache
NextJS as of today does not support configuration of .next/cache,
where temporary files, routes and images are cached for faster
execution.

Since in production the app is run from read-only filesystems, having
the cache in a write-fs is crucial. So we symlink our way to the
solution

Similar, nftdata is mounted.

Note that .env.local allows to specify the STORAGE_DIR, but in fact
this is no real option, if users do not at the same time adjust
Dockerfile, docker-compose etc. So better leave it untouched!