Production iteration (breaking schema change) (#41)
* Remove fsdemo endpoint

* Use config STORAGE_DIR in asset endpoint

The STORAGE_DIR/upload  directory is used for uploading images and videos and will
therefore contain several large files, at least temporarily.

We avoid using the system's os.tmpdir(), as this may be exploited by
Arbitrary Code Execution (node building also uses the same directory)

* Generalize asset-upload (#40)

This adapts the asset-endpoint to support all assetTypes, with the
NFT image upload being a common case

* Define /srv/data Mountpoint for nftdata and .next/cache

NextJS as of today does not support configuration of .next/cache,
where temporary files, routes and images are cached for faster
execution.

Since in production the app is run from read-only filesystems, having
the cache in a write-fs is crucial. So we symlink our way to the
solution

Similar, nftdata is mounted.

Note that .env.local allows to specify the STORAGE_DIR, but in fact
this is no real option, if users do not at the same time adjust
Dockerfile, docker-compose etc. So better leave it untouched!

* Update docker-compose.yml to proper volume mount point

* Add stub for next-cache

* Asset: Rename file_name->file_path, contract-directory creation atomic