Use NFTCard for Collection DefaultNFT
The Default-NFT is a "full" NFT as well, therefore it shall be shown
as a NFTCard, which forwards to the NFTView.

Since the DefaultNFT does not have a SLID (and also no real anchor),
and consequently can also not be owned directly,
adaptions to NFTCard and NFTView were needed to show proper captions.

Went with a saticCaption-Approach for more flexibility. A staticCapture
is enforced for NFTs with the anchor 'default'