Baseline, assets join table and soft-delete (#60)
* all: move prisma into it's own folder

* all: format prisma schema

* all: add second baseline (i.e.: nft <-> assets)

* schema: create assets_nfts join table

This table will be responsible for holding the relationship between a
specific NFT and its assets. It'll also be responsible for storing the
history of assets that a given NFT has had in the past.

It should replace Prisma's implicit join table in future patches

* docker: remove redundant line copying schema from Dockerfile

Since schema is now inside the prisma folder

* all: always run migrations instead of force pushing the db

* schema: drop implicit nft <-> assets join table

The AssetsNFTs table replaces it, since it allows more information to be
stored alongside the relationship.

* studio: load nested asset from assets nfts

* api/v1: adapt collection endpoint to match new asset to nft structure

* schema: adds metadata information on assets nfts relationship

Due to how difficult are *data migrations* on Prisma, the `assignedBy`
relationship is marked as optional. This is not optimal, but it is a
temporary measure, since that this could lead to invalid data on the
table. Thus, after a data migration, we'll mark the field as obligatory

* schema: drop pluralized version from AssetNFT model

* assets: detect nfts primary asset using a tiny service

* api/internal: upgrade asset uploading logic to match new assets nfts relationship

* api/internal: touch asset nft model when soft deleting it