studio/edit: Treat image as regular asset
Images are assets with the tag "image". In terms of NFT and metadata,
the "image" tag is equal to any other asset with "my-arbitrary-key".
Only marketplaces and applications preferably look for an image tag THE
representative meta-datum.

This logic is also in studio by the discoverPrimaryAsset() function,
however, treating images differently does not make sense from an
NFT-Editing perspective